fix: change receiver_memo type from i128 to u32 in Escrow struct

+++ contracts/escrow/src/storage/types.rs
@@ -12,7 +12,7 @@ pub struct Escrow {
     pub milestones: Vec<Milestone>,
     pub flags: Flags,
     pub trustline: Trustline,
-    pub receiver_memo: i128,
+    pub receiver_memo: u32,
 }
 
 #[contracttype]
